max round

+++ src/matOptimize/transpose_vcf/transposed_vcf_patch.cpp
@@ -181,7 +181,7 @@ struct output_vcf_rows {
         size_t idx=start_idx;
         std::vector<mut_iterator> iters;
         for (const auto &samp : all_samples) {
-            if (samp.bfs_idx!=-1) {                
+            if (samp.bfs_idx!=(size_t)-1) {
                 iters.emplace_back(samp, positions[idx]);
             }
         }
